+++ demisto_sdk/commands/common/hook_validations/wizard.py
@@ -25,9 +25,10 @@ class WizardValidator(ContentEntityValidator):
             return True
 
         deps_are_valid = True
-        wizard_dependencies = {}
-        for category in self.current_file.get("dependency_packs", {}).values():
-            wizard_dependencies.update(category.get("packs", []))
+        wizard_dependencies = set()
+        for pack_metadata in self.current_file.get("dependency_packs", []):
+            packs = [pack.get('name', '') for pack in pack_metadata.get("packs", [])]
+            wizard_dependencies.update(packs)
         packs = id_set_file.get('Packs', {})
         for dep in wizard_dependencies:
             if dep not in packs: