+++ demisto_sdk/commands/common/tests/structure_test.py
@@ -318,6 +318,27 @@ class TestStructureValidator:
         captured = capsys.readouterr().out
         assert f'Missing the field "{missing_field}" in root' in captured
 
+    @pytest.mark.parametrize('missing_field',
+                             ('dependency_packs', 'wizard', 'name', 'id', 'fromVersion'))
+    def test_wizard_missing_field(self, repo, capsys, missing_field: str):
+        """
+        Given
+                A Job object in a repo, with one of the required fields missing
+        When
+                Validating the file
+        Then
+                Ensure the structure validator raises a suitable error
+        """
+        pack = repo.create_pack()
+        wizard = pack.create_wizard(name='wizard_name')
+        wizard.remove(missing_field)
+
+        validator = StructureValidator(wizard.path, is_new_file=True)
+        with ChangeCWD(repo.path):
+            assert not validator.is_valid_file()
+        captured = capsys.readouterr().out
+        assert f'Missing the field "{missing_field}" in root' in captured
+
     def test_validate_field_with_aliases__valid(self, pack: Pack):
         """
         Given